requirements.txt
# SPADE Framework para agentes
# Nota: SPADE 4.1.0 com slixmpp padrão (não multiplatform)
spade==4.1.0
slixmpp==1.9.1

# SUMO Integration
traci>=1.14.0
sumolib>=1.14.0

# Utilidades
numpy>=1.21.0
python-dotenv>=0.19.0

# Async support
aiohttp>=3.8.0

# Para logging e debugging
coloredlogs>=15.0

# Workaround para Python 3.14: usar variável de ambiente
# Caso ainda haja erro, use: PYO3_USE_ABI3_FORWARD_COMPATIBILITY=1 pip install -r requirements.txt
